requirements.txt
streamlit
pandas
st-gsheets-connection
pyzbar
pillow